+++ src/service/ExtendedParams.ts
@@ -3,12 +3,9 @@
  *  Licensed under the MIT License. See LICENSE in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { CompletionParams, TextDocumentIdentifier, TextDocumentPositionParams } from 'vscode-languageserver';
-import { ComposeDocument } from './ComposeDocument';
-
-export interface TextDocumentParams { // This interface ought to exist in `vscode-languageserver`, like `TextDocumentPositionParams`, but here we are...
-    textDocument: TextDocumentIdentifier;
-}
+import type { CompletionParams, TextDocumentPositionParams } from 'vscode-languageserver';
+import type { ComposeDocument } from './ComposeDocument';
+import type { TextDocumentParams } from '../common/TextDocumentParams';
 
 export interface ExtendedParams extends TextDocumentParams {
     document: ComposeDocument;