+++ src/test/ComposeDocument.test.ts
@@ -5,8 +5,8 @@
 
 import { expect } from 'chai';
 import { InitializeParams, Position, TextDocuments } from 'vscode-languageserver';
-import type { ComposeLanguageClientCapabilities } from '../client/ClientCapabilities';
-import { DocumentSettings, DocumentSettingsNotification, DocumentSettingsNotificationParams, DocumentSettingsRequest, LF } from '../client/DocumentSettings';
+import type { ComposeLanguageClientCapabilities } from '../common/ComposeLanguageClientCapabilities';
+import { DocumentSettings, DocumentSettingsNotification, DocumentSettingsNotificationParams, DocumentSettingsRequest, LF } from '../common/DocumentSettingsClientCapabilities';
 import { ComposeDocument } from '../service/ComposeDocument';
 import { ExtendedPositionParams, PositionInfo } from '../service/ExtendedParams';
 import { runWithContext } from '../service/utils/ActionContext';
@@ -58,7 +58,7 @@ describe('ComposeDocument', () => {
         testConnection = new TestConnection(initParams);
 
         // Set up a client listener to respond to doc settings requests
-        testConnection.client.onRequest(DocumentSettingsRequest.method, (params) => {
+        testConnection.client.onRequest(DocumentSettingsRequest.type, (params) => {
             return {
                 eol: LF,
                 tabSize: 4,
@@ -655,7 +655,7 @@ describe('ComposeDocument', () => {
                 (sharedComposeDocument as any).documentSettings = undefined;
 
                 const requestPromise = new Promise<void>((resolve, reject) => {
-                    testConnection.client.onRequest(DocumentSettingsRequest.method, (params) => {
+                    testConnection.client.onRequest(DocumentSettingsRequest.type, (params) => {
                         resolve();
                         return {
                             eol: LF,