+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.5.0 - 9 February 2026
+### Breaking Changes
+* `TelemetryEvent`, `AlternateYamlLanguageServiceClientCapabilities`, and `DocumentSettings*` types have moved from `lib/client/*` to `@microsoft/compose-language-service/client`.
+
+### Added
+* An entrypoint was added for `client`. This is intended for use by any client to help with interacting with this language service.
+* An entrypoint was added for `vscode`. This is intended for use by VS Code in particular to help with interacting with this language service.
+* An ESM build of the entire package is now included, in addition to the prior CJS build.
+
 ## 0.4.0 - 17 April 2025
 ### Breaking Changes
 * The service startup CodeLens has changed the command it will call to `vscode-containers.compose.up`.